Move the rest of the e2e commands to new format
