+++ e2e-2/index.js
@@ -8,9 +8,23 @@ const output = execSync(`npx jest e2e-2 --listTests`);
 async function main() {
     let failed = false;
 
+    let filterText;
+
+    if (process.argv.includes('--filter')) {
+        const index = process.argv.indexOf('--filter') + 1;
+        filterText = process.argv[index];
+    }
+
     const testFiles = String(output)
         .split('\n')
-        .filter(line => line !== '');
+        .filter(line => line !== '')
+        .filter(line => {
+            if (filterText) {
+                return line.includes(filterText);
+            } else {
+                return line;
+            }
+        });
 
     for (let i = 0; i < testFiles.length; i++) {
         const testClient = new Docker();
