+++ e2e-2/init/errors.test.ts
@@ -1,0 +1,16 @@
+import colors from 'colors';
+
+import { assembleOutput, run } from '../../ts-e2e/utils';
+
+colors.enable();
+
+test('errors when in invalid state', () => {
+    run('rm -rf .git');
+    const [out, exitCode] = run('guet init');
+    const expectedOutput = assembleOutput([
+        'git not installed in this directory.'.red,
+    ]);
+
+    expect(out).toEqual(expectedOutput);
+    expect(exitCode).toEqual(1);
+});
