+++ e2e-2/set/errors.test.ts
@@ -1,0 +1,18 @@
+import colors from 'colors';
+
+import { assembleOutput, run } from '../../ts-e2e/utils';
+
+colors.enable();
+
+test('should display error message when given unknown initials', () => {
+    run('git init');
+    run('guet init');
+    run('guet add fn "first name" fn@example.com');
+    const [out, errorCode] = run('guet set fn fn2');
+
+    expect(out).toEqual(
+        assembleOutput(['No committer exists with the initials "fn2".'.red])
+    );
+
+    expect(errorCode).toEqual(1);
+});
