+++ jest.setup.ts
@@ -1,3 +1,6 @@
+import colors from 'colors';
+
 beforeEach(() => {
+    colors.enable();
     jest.resetAllMocks();
 });
