+++ ts-e2e/utils.ts
@@ -6,7 +6,7 @@ import { configPath, getGitPath } from '../src/utils';
 export function run(command: string): [string, number] {
     try {
         const output = execSync(command, { encoding: 'utf8' });
-        return [output, 0];
+        return [String(output), 0];
     } catch (e: any) {
         return [String(e.stdout), 1];
     }
